# apps/api-client-go/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@boxlite.com
    name: BoxLite Platforms Inc.
    url: https://www.boxlite.io
  description: BoxLite AI platform API Docs
  title: BoxLite
  version: "1.0"
servers:
- url: http://localhost:3000
paths:
...
  /box/{boxIdOrName}/recover:
    post:
      operationId: recoverBox
      parameters:
      - description: Use with JWT to specify the organization ID
        explode: false
        in: header
        name: X-BoxLite-Organization-ID
        required: false
        schema:
          type: string
        style: simple
      - description: ID or name of the box
        explode: false
        in: path
        name: boxIdOrName
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Box"
          description: Recovery initiated
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Recover box from error state
      tags:
      - box
...
  /runners/{id}:
    delete:
      operationId: deleteRunner
      parameters:
      - description: Runner ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Use with JWT to specify the organization ID
        explode: false
        in: header
        name: X-BoxLite-Organization-ID
        required: false
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Delete runner
      tags:
      - runners
    get:
      operationId: getRunnerById
      parameters:
      - description: Runner ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Use with JWT to specify the organization ID
        explode: false
        in: header
        name: X-BoxLite-Organization-ID
        required: false
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Runner"
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Get runner by ID
      tags:
      - runners
...
  /runners/{id}/scheduling:
    patch:
      operationId: updateRunnerScheduling
      parameters:
      - description: Runner ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Use with JWT to specify the organization ID
        explode: false
        in: header
        name: X-BoxLite-Organization-ID
        required: false
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Runner"
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Update runner scheduling status
      tags:
      - runners
...
  /admin/runners:
    get:
      operationId: adminListRunners
      parameters:
      - description: Filter runners by region ID
        explode: true
        in: query
        name: regionId
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/AdminRunner"
                type: array
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: List all runners
      tags:
      - admin
    post:
      operationId: adminCreateRunner
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateRunner"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRunnerResponse"
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Create runner
      tags:
      - admin
  /admin/runners/{id}:
    delete:
      operationId: adminDeleteRunner
      parameters:
      - description: Runner ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Delete runner
      tags:
      - admin
    get:
      operationId: adminGetRunnerById
      parameters:
      - description: Runner ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRunner"
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Get runner by ID
      tags:
      - admin
  /admin/runners/{id}/scheduling:
    patch:
      operationId: adminUpdateRunnerScheduling
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Update runner scheduling status
      tags:
      - admin
  /admin/box/{boxId}/recover:
    post:
      operationId: adminRecoverBox
      parameters:
      - description: ID of the box
        explode: false
        in: path
        name: boxId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Box"
          description: Recovery initiated
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Recover box from error state as an admin
      tags:
      - admin
...
components:
  schemas:
...
    RegionType:
      description: The type of the region
      enum:
      - shared
      - dedicated
      - custom
      type: string
...
    BoxState:
      description: The state of the box
      enum:
      - creating
      - restoring
      - destroyed
      - destroying
      - started
      - stopped
      - starting
      - stopping
      - error
      - unknown
      - archived
      - archiving
      - resizing
      type: string
    BoxDesiredState:
      description: The desired state of the box
      enum:
      - destroyed
      - started
      - stopped
      - resized
      type: string
    BoxVolume:
      example:
        volumeId: volume123
        mountPath: /data
        subpath: users/alice
      properties:
        volumeId:
          description: The ID of the volume
          example: volume123
          type: string
        mountPath:
          description: The mount path for the volume
          example: /data
          type: string
        subpath:
          description: "Optional subpath within the volume to mount. When specified,\
            \ only this S3 prefix will be accessible. When omitted, the entire volume\
            \ is mounted."
          example: users/alice
          type: string
      required:
      - mountPath
      - volumeId
      type: object
    Box:
      example:
        id: aB3cD4eF5gH6
        organizationId: organization123
        name: MyBox
        user: boxlite
        env:
          NODE_ENV: production
        labels:
          boxlite.io/public: "true"
        public: false
        networkBlockAll: false
        networkAllowList: "192.168.1.0/16,10.0.0.0/24"
        target: local
        image: boxlite/base
        cpu: 2
        gpu: 0
        memory: 4
        disk: 10
        state: creating
        desiredState: destroyed
        errorReason: The box is not running
        recoverable: true
        autoStopInterval: 30
        autoDeleteInterval: 30
        volumes:
        - volumeId: volume123
          mountPath: /data
          subpath: users/alice
        - volumeId: volume123
          mountPath: /data
          subpath: users/alice
        createdAt: 2024-10-01T12:00:00Z
        updatedAt: 2024-10-01T12:00:00Z
        class: small
        daemonVersion: 1.0.0
        runnerId: runner123
        toolboxProxyUrl: https://proxy.app.boxlite.io/toolbox
      properties:
        id:
          description: The public 12-character Box ID
          example: aB3cD4eF5gH6
          type: string
        organizationId:
          description: The organization ID of the box
          example: organization123
          type: string
        name:
          description: The name of the box
          example: MyBox
          type: string
        user:
          description: The user associated with the project
          example: boxlite
          type: string
        env:
          additionalProperties:
            type: string
          description: Environment variables for the box
          example:
            NODE_ENV: production
          type: object
        labels:
          additionalProperties:
            type: string
          description: Labels for the box
          example:
            boxlite.io/public: "true"
          type: object
        public:
          description: Whether the box http preview is public
          example: false
          type: boolean
        networkBlockAll:
          description: Whether to block all network access for the box
          example: false
          type: boolean
        networkAllowList:
          description: Comma-separated list of allowed CIDR network addresses for
            the box
          example: "192.168.1.0/16,10.0.0.0/24"
          type: string
        target:
          description: The target environment for the box
          example: local
          type: string
        image:
          description: The image used for the box
          example: boxlite/base
          type: string
        cpu:
          description: The CPU quota for the box
          example: 2
          type: number
        gpu:
          description: The GPU quota for the box
          example: 0
          type: number
        memory:
          description: The memory quota for the box
          example: 4
          type: number
        disk:
          description: The disk quota for the box
          example: 10
          type: number
        state:
          allOf:
          - $ref: "#/components/schemas/BoxState"
          description: The state of the box
          example: creating
        desiredState:
          allOf:
          - $ref: "#/components/schemas/BoxDesiredState"
          description: The desired state of the box
          example: destroyed
        errorReason:
          description: The error reason of the box
          example: The box is not running
          type: string
        recoverable:
          description: Whether the box error is recoverable.
          example: true
          type: boolean
        autoStopInterval:
          description: Auto-stop interval in minutes (0 means disabled)
          example: 30
          type: number
        autoDeleteInterval:
          description: "Auto-delete interval in minutes (negative value means disabled,\
            \ 0 means delete immediately upon stopping)"
          example: 30
          type: number
        volumes:
          description: Array of volumes attached to the box
          items:
            $ref: "#/components/schemas/BoxVolume"
          type: array
        createdAt:
          description: The creation timestamp of the box
          example: 2024-10-01T12:00:00Z
          type: string
        updatedAt:
          description: The last update timestamp of the box
          example: 2024-10-01T12:00:00Z
          type: string
        class:
          deprecated: true
          description: The class of the box
          enum:
          - small
          - medium
          - large
          example: small
          type: string
        daemonVersion:
          description: The version of the daemon running in the box
          example: 1.0.0
          type: string
        runnerId:
          description: The runner ID of the box
          example: runner123
          type: string
        toolboxProxyUrl:
          description: The toolbox proxy URL for the box
          example: https://proxy.app.boxlite.io/toolbox
          type: string
      required:
      - cpu
      - disk
      - env
      - gpu
      - id
      - labels
      - memory
      - name
      - networkBlockAll
      - organizationId
      - public
      - target
      - toolboxProxyUrl
      - user
      type: object
...
    CreateRunnerResponse:
      example:
        id: runner123
        apiKey: blk_svc_1234567890
      properties:
        id:
          description: The ID of the runner
          example: runner123
          type: string
        apiKey:
          description: The API key for the runner
          example: blk_svc_1234567890
          type: string
      required:
      - apiKey
      - id
      type: object
    BoxClass:
      description: The class of the runner
      enum:
      - small
      - medium
      - large
      type: string
    RunnerState:
      description: The state of the runner
      enum:
      - initializing
      - ready
      - disabled
      - decommissioned
      - unresponsive
      type: string
...
    Runner:
      example:
        id: runner123
        domain: runner1.example.com
        apiUrl: https://api.runner1.example.com
        proxyUrl: https://proxy.runner1.example.com
        cpu: 8
        memory: 16
        disk: 100
        gpu: 1
        gpuType: gpuType
        class: small
        currentCpuUsagePercentage: 45.6
        currentMemoryUsagePercentage: 68.2
        currentDiskUsagePercentage: 33.8
        currentAllocatedCpu: 4000
        currentAllocatedMemoryGiB: 8000
        currentAllocatedDiskGiB: 50000
        currentStartedBoxes: 5
        availabilityScore: 85
        region: us
        name: runner1
        state: initializing
        lastChecked: 2024-10-01T12:00:00Z
        unschedulable: false
        createdAt: 2023-10-01T12:00:00Z
        updatedAt: 2023-10-01T12:00:00Z
        version: "0"
        apiVersion: "0"
        appVersion: v0.0.0-dev
      properties:
        id:
          description: The ID of the runner
          example: runner123
          type: string
        domain:
          description: The domain of the runner
          example: runner1.example.com
          type: string
        apiUrl:
          description: The API URL of the runner
          example: https://api.runner1.example.com
          type: string
        proxyUrl:
          description: The proxy URL of the runner
          example: https://proxy.runner1.example.com
          type: string
        cpu:
          description: The CPU capacity of the runner
          example: 8
          type: number
        memory:
          description: The memory capacity of the runner in GiB
          example: 16
          type: number
        disk:
          description: The disk capacity of the runner in GiB
          example: 100
          type: number
        gpu:
          description: The GPU capacity of the runner
          example: 1
          type: number
        gpuType:
          description: The type of GPU
          type: string
        class:
          allOf:
          - $ref: "#/components/schemas/BoxClass"
          description: The class of the runner
          example: small
        currentCpuUsagePercentage:
          description: Current CPU usage percentage
          example: 45.6
          type: number
        currentMemoryUsagePercentage:
          description: Current RAM usage percentage
          example: 68.2
          type: number
        currentDiskUsagePercentage:
          description: Current disk usage percentage
          example: 33.8
          type: number
        currentAllocatedCpu:
          description: Current allocated CPU
          example: 4000
          type: number
        currentAllocatedMemoryGiB:
          description: Current allocated memory in GiB
          example: 8000
          type: number
        currentAllocatedDiskGiB:
          description: Current allocated disk in GiB
          example: 50000
          type: number
        currentStartedBoxes:
          description: Current number of started boxes
          example: 5
          type: number
        availabilityScore:
          description: Runner availability score
          example: 85
          type: number
        region:
          description: The region of the runner
          example: us
          type: string
        name:
          description: The name of the runner
          example: runner1
          type: string
        state:
          allOf:
          - $ref: "#/components/schemas/RunnerState"
          description: The state of the runner
          example: initializing
        lastChecked:
          description: The last time the runner was checked
          example: 2024-10-01T12:00:00Z
          type: string
        unschedulable:
          description: Whether the runner is unschedulable
          example: false
          type: boolean
        createdAt:
          description: The creation timestamp of the runner
          example: 2023-10-01T12:00:00Z
          type: string
        updatedAt:
          description: The last update timestamp of the runner
          example: 2023-10-01T12:00:00Z
          type: string
        version:
          deprecated: true
          description: The version of the runner (deprecated in favor of apiVersion)
          example: "0"
          type: string
        apiVersion:
          deprecated: true
          description: The api version of the runner
          example: "0"
          type: string
        appVersion:
          deprecated: true
          description: The app version of the runner
          example: v0.0.0-dev
          type: string
      required:
      - apiVersion
      - class
      - cpu
      - createdAt
      - disk
      - id
      - memory
      - name
      - region
      - state
      - unschedulable
      - updatedAt
      - version
      type: object
...
    AdminCreateRunner:
      example:
        regionId: regionId
        name: name
        apiKey: apiKey
        apiVersion: "2"
        domain: runner1.example.com
        apiUrl: https://api.runner1.example.com
        proxyUrl: https://proxy.runner1.example.com
        cpu: 8
        memoryGiB: 16
        diskGiB: 100
      properties:
        regionId:
          type: string
        name:
          type: string
        apiKey:
          type: string
        apiVersion:
          description: The api version of the runner to create
          example: "2"
          pattern: ^(0|2)$
          type: string
        domain:
          description: The domain of the runner
          example: runner1.example.com
          type: string
        apiUrl:
          description: The API URL of the runner
          example: https://api.runner1.example.com
          type: string
        proxyUrl:
          description: The proxy URL of the runner
          example: https://proxy.runner1.example.com
          type: string
        cpu:
          description: The CPU capacity of the runner
          example: 8
          type: number
        memoryGiB:
          description: The memory capacity of the runner in GiB
          example: 16
          type: number
        diskGiB:
          description: The disk capacity of the runner in GiB
          example: 100
          type: number
      required:
      - apiKey
      - apiVersion
      - name
      - regionId
      type: object
    AdminRunner:
      example:
        id: runner123
        domain: runner1.example.com
        apiUrl: https://api.runner1.example.com
        proxyUrl: https://proxy.runner1.example.com
        cpu: 8
        memory: 16
        disk: 100
        gpu: 1
        gpuType: gpuType
        class: small
        currentCpuUsagePercentage: 45.6
        currentMemoryUsagePercentage: 68.2
        currentDiskUsagePercentage: 33.8
        currentAllocatedCpu: 4000
        currentAllocatedMemoryGiB: 8000
        currentAllocatedDiskGiB: 50000
        currentStartedBoxes: 5
        availabilityScore: 85
        region: us
        name: runner1
        state: initializing
        lastChecked: 2024-10-01T12:00:00Z
        unschedulable: false
        createdAt: 2023-10-01T12:00:00Z
        updatedAt: 2023-10-01T12:00:00Z
        version: "0"
        apiVersion: "0"
        appVersion: v0.0.0-dev
        regionType: shared
      properties:
        id:
          description: The ID of the runner
          example: runner123
          type: string
        domain:
          description: The domain of the runner
          example: runner1.example.com
          type: string
        apiUrl:
          description: The API URL of the runner
          example: https://api.runner1.example.com
          type: string
        proxyUrl:
          description: The proxy URL of the runner
          example: https://proxy.runner1.example.com
          type: string
        cpu:
          description: The CPU capacity of the runner
          example: 8
          type: number
        memory:
          description: The memory capacity of the runner in GiB
          example: 16
          type: number
        disk:
          description: The disk capacity of the runner in GiB
          example: 100
          type: number
        gpu:
          description: The GPU capacity of the runner
          example: 1
          type: number
        gpuType:
          description: The type of GPU
          type: string
        class:
          allOf:
          - $ref: "#/components/schemas/BoxClass"
          description: The class of the runner
          example: small
        currentCpuUsagePercentage:
          description: Current CPU usage percentage
          example: 45.6
          type: number
        currentMemoryUsagePercentage:
          description: Current RAM usage percentage
          example: 68.2
          type: number
        currentDiskUsagePercentage:
          description: Current disk usage percentage
          example: 33.8
          type: number
        currentAllocatedCpu:
          description: Current allocated CPU
          example: 4000
          type: number
        currentAllocatedMemoryGiB:
          description: Current allocated memory in GiB
          example: 8000
          type: number
        currentAllocatedDiskGiB:
          description: Current allocated disk in GiB
          example: 50000
          type: number
        currentStartedBoxes:
          description: Current number of started boxes
          example: 5
          type: number
        availabilityScore:
          description: Runner availability score
          example: 85
          type: number
        region:
          description: The region of the runner
          example: us
          type: string
        name:
          description: The name of the runner
          example: runner1
          type: string
        state:
          allOf:
          - $ref: "#/components/schemas/RunnerState"
          description: The state of the runner
          example: initializing
        lastChecked:
          description: The last time the runner was checked
          example: 2024-10-01T12:00:00Z
          type: string
        unschedulable:
          description: Whether the runner is unschedulable
          example: false
          type: boolean
        createdAt:
          description: The creation timestamp of the runner
          example: 2023-10-01T12:00:00Z
          type: string
        updatedAt:
          description: The last update timestamp of the runner
          example: 2023-10-01T12:00:00Z
          type: string
        version:
          deprecated: true
          description: The version of the runner (deprecated in favor of apiVersion)
          example: "0"
          type: string
        apiVersion:
          deprecated: true
          description: The api version of the runner
          example: "0"
          type: string
        appVersion:
          deprecated: true
          description: The app version of the runner
          example: v0.0.0-dev
          type: string
        regionType:
          allOf:
          - $ref: "#/components/schemas/RegionType"
          description: The region type of the runner
          example: shared
      required:
      - apiVersion
      - class
      - cpu
      - createdAt
      - disk
      - id
      - memory
      - name
      - region
      - state
      - unschedulable
      - updatedAt
      - version
      type: object
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      description: API Key access
      scheme: bearer
      type: http
    oauth2:
      openIdConnectUrl: http://localhost:3000/.well-known/openid-configuration
      type: openIdConnect